+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # v 2.10.0 (?)
 Changes in this release:
 - Add support for custom `agent_map` enabling testing or remote io scenarios.
-- Make sure that the resource is serialized when `deploy_all` is used
+- Fix resource serialization (#437)
 
 # v 2.9.0 (2023-11-29)
 Changes in this release: